Fix a ton of bugs. Get all tests to pass.

+++ lark_template.rb
@@ -263,7 +263,7 @@ compass_css_framework = "blueprint" if compass_css_framework.nil?
 design = template_options["design"].nil? ? ask("Which design? none (default), bluetrip, compass").downcase : template_options["design"]
 design = "none" if design.nil?
 
-require_activation = (template_options["require_activation"] == "true")
+require_activation = (template_options["require_activation"].to_s == "true")
 
 smtp_address = template_options["smtp_address"]
 smtp_domain = template_options["smtp_domain"]
@@ -348,6 +348,9 @@ in_root do
 end
 environment 'config.middleware.use "Rack::Bug"', :env => 'development'
 environment 'config.middleware.use "Rack::Bug"', :env => 'staging'
+
+environment 'config.action_mailer.delivery_method = :smtp', :env => 'production'
+environment 'config.action_mailer.delivery_method = :smtp', :env => 'staging'
 
 commit_state "Set up staging environment and hooked up Rack::Bug"
 
@@ -484,6 +487,12 @@ end
 file 'config/database.yml', load_pattern("config/database.#{database}.yml", 'default', binding)
 file 'db/populate/01_sample_seed.rb', load_pattern('db/populate/01_sample_seed.rb')
 
+if require_activation
+  account_create_flash = "Your account has been created. Please check your e-mail for your account activation instructions."
+else
+  account_create_flash = "Account registered!"
+end
+
 # locale
 file 'config/locales/en.yml', load_pattern('config/locales/en.yml', 'default', binding)
 
@@ -553,13 +562,11 @@ end
 file 'test/functional/pages_controller_test.rb', load_pattern('test/functional/pages_controller_test.rb', 'default', binding)
 file 'test/functional/password_resets_controller_tests.rb', load_pattern('test/functional/password_resets_controller_tests.rb')
 
+new_user_contained_text = 'I18n.t("flash.accounts.create.notice")'
+
 new_user_extra_fields = ""
-new_user_contained_text = ""
-if require_activation
-  new_user_contained_text = 'Your account has been created'
-else
+unless require_activation
   new_user_extra_fields = load_snippet('new_user_extra_fields')
-  new_user_contained_text = 'Account registered!'
 end
 
 file 'test/integration/new_user_can_register_test.rb', load_pattern('test/integration/new_user_can_register_test.rb', 'default', binding)
@@ -702,7 +709,7 @@ file 'config/routes.rb', load_pattern('config/routes.rb', 'default', binding)
 commit_state "routing"
 
 # optionally convert html/erb/css to haml/sass
-if template_engine == 'haml' || design == "compass"
+if template_engine == 'haml'
   def get_indent(line)
     line = line.to_s
     space_areas = line.scan(/^\s+/)
@@ -710,36 +717,65 @@ if template_engine == 'haml' || design == "compass"
   end
 
   def block_start?(line)
-    block_starters = [/\s+do\W/, /^-\s+while/, /^-\s+module/, /^-\s+begin/,
+    block_starters = [/\s+do/, /^-\s+while/, /^-\s+module/, /^-\s+begin/,
                       /^-\s+case/, /^-\s+class/, /^-\s+unless/, /^-\s+for/, 
-                      /^-\s+until/, /^-\s+if/]
+                      /^-\s+until/, /^-\s*if/]
 
     line = line.to_s
-    line.strip =~ /^-/ && block_starters.any?{|bs| line =~ bs}
+    line.strip =~ /^-/ && block_starters.any?{|bs| line.strip =~ bs}
   end
 
   def block_end?(line)
     line = line.to_s
     line.strip =~ /^-\send$/
   end
-  
+
   def ie_block_start?(line)
     line = line.to_s
-    line =~ /\[if IE\]>/i
+    line =~ /\[if/i && line =~ /IE/ && line.strip =~ /\]>$/
   end
-  
+
   def ie_block_end?(line)
     line = line.to_s
     line =~ /<!\[endif\]/i
   end
 
+  def comment_line?(line)
+    line = line.to_s
+    line.strip =~ /^\//
+  end
+
   def indent(line, steps = 0)
+    line = line.to_s
     exceptions = [/\s+else\W/, /^-\s+elsif/, /^-\s+when/, /^-\s+ensure/, /^-\s+rescue/]
-    return if exceptions.any?{|ex| line =~ ex}
+    return if exceptions.any?{|ex| line.strip =~ ex}
 
     steps ||= 0
     line = line.to_s
     ("  " * steps) + line
+  end
+
+  def alter_lines(lines, altered_lines)
+    altered_lines.each do |pair|
+      line_number, text = pair
+      lines[line_number] = text
+    end
+    lines
+  end
+
+  def indent_lines(lines, indented_lines)
+    indented_lines.each do |pair|
+      line_number, indent_by = pair
+      lines[line_number] = indent(lines[line_number], indent_by)
+    end
+    lines
+  end
+
+  def remove_lines(lines, goner_lines)
+    goner_lines.each do |i|
+      lines[i] = nil
+    end
+    lines.compact
   end
   
   in_root do
@@ -756,18 +792,18 @@ if template_engine == 'haml' || design == "compass"
       altered_lines = []
       inside_ie_block = false
       just_passed_ie_block = false
-      
+
       lines.reverse_each do |line|
         line_number -= 1
-        
+
         if just_passed_ie_block
-          altered_lines << [line_number, line.sub('/', '/[if IE]')]
+          altered_lines << [line_number, line.sub('/', "/" + just_passed_ie_block)]
           just_passed_ie_block = false
         elsif ie_block_start?(line)
           goner_lines << line_number
           inside_ie_block = false
           stack.pop
-          just_passed_ie_block = true
+          just_passed_ie_block = line.strip.chop
         elsif ie_block_end?(line)
           goner_lines << line_number
           inside_ie_block = true
@@ -775,35 +811,47 @@ if template_engine == 'haml' || design == "compass"
         elsif inside_ie_block
           match = line.match(/<haml[^>]*>([^<]+)<\/haml/)
           string = match && match[1]
-          fixed_ie = string && "= #{CGI::unescapeHTML(CGI::unescapeHTML(string.strip))}\n"
-          altered_lines << [line_number, fixed_ie]
+          string = string ? "= #{CGI::unescapeHTML(CGI::unescapeHTML(string.strip))}\n" : "#{line.strip}\n"
+          altered_lines << [line_number, string]
           indented_lines << [line_number, stack.last]
-          
         elsif block_end?(line)
           stack << 1
           goner_lines << line_number
-          
         elsif block_start?(line)
           stack.pop
-          
         else
           indented_lines << [line_number, stack.last]
         end
       end
-      
-      altered_lines.each do |pair|
-        line_number, text = pair
-        lines[line_number] = text
-      end
-      
-      indented_lines.each do |pair|
-        line_number, indent_by = pair
-        lines[line_number] = indent(lines[line_number], indent_by)
+
+      lines = alter_lines(lines, altered_lines)
+      lines = indent_lines(lines, indented_lines)
+      lines = remove_lines(lines, goner_lines)
+
+      altered_lines = []
+      indented_lines = [] 
+      goner_lines = []
+
+      line_number = -1
+      lines.each_cons(3) do |three_lines|
+        line_number += 1
+        line2_number = line_number + 1
+        middle_indented_by_one = (get_indent(three_lines[1]) - get_indent(three_lines[0]) == 1)
+        top_indented_more = (get_indent(three_lines[0]) >= get_indent(three_lines[2]))
+        commented = (three_lines[0].to_s.strip =~ /^\//)
+
+        if(top_indented_more && middle_indented_by_one && !commented)
+          if (three_lines[1].strip =~ /^=/)
+            altered_lines << [line_number, (three_lines[0].rstrip + three_lines[1].lstrip)]
+          else
+            altered_lines << [line_number, (three_lines[0].rstrip + " " + three_lines[1].lstrip)]
+          end
+          goner_lines << line2_number
+        end
       end
 
-      goner_lines.each do |i|
-        lines.delete_at(i)
-      end
+      lines = alter_lines(lines, altered_lines)
+      lines = remove_lines(lines, goner_lines)
       
       File.open(destination, "w") do |f|
         f.write(lines.join)
@@ -811,11 +859,13 @@ if template_engine == 'haml' || design == "compass"
       
       File.delete(origin)
     end
-    
-    Dir["public/stylesheets/**/*.css"].each do |file|
-      run "css2sass #{file} #{file.gsub(/\.css$/, '.sass')}"
-      File.delete(file)
-    end
+  end
+end
+
+if template_engine == "haml" || design == "compass"
+  Dir["public/stylesheets/**/*.css"].each do |file|
+    run "css2sass #{file} #{file.gsub(/\.css$/, '.sass')}"
+    File.delete(file)
   end
 end
